+++ backend/src/modules/user/user.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
@@ -24,21 +25,27 @@ import {
   CreateUserUploadedFiles,
   UpdateUserUploadedFiles,
 } from 'src/modules/user/types/user.types';
+import { Auth } from 'src/core/decorators/auth.decorator';
+import { JwtAuthGuard } from 'src/modules/auth/guards/jwt-auth.guard';
+import { AuthenticatedUser } from 'src/core/decorators/authenticated-user.decorator';
 
 @Controller(Routes.Users)
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
+  @Auth(JwtAuthGuard)
   @Get()
   public async findAll(@Query() query?: string) {
     return this.userService.findAll(deserializeQueryString(query));
   }
 
+  @Auth(JwtAuthGuard)
   @Get(':id')
   public async findOneById(@Param('id') id: UserPublicEntity['id'], @Query() query?: string) {
     return this.userService.findOne(_.merge(deserializeQueryString(query), { where: { id } }));
   }
 
+  @Auth(JwtAuthGuard)
   @Post()
   @UseInterceptors(FileFieldsInterceptor([{ name: 'image', maxCount: 1 }]))
   public async create(
@@ -57,11 +64,13 @@ export class UserController {
     return this.userService.create(createUserDto, files);
   }
 
+  @Auth(JwtAuthGuard)
   @Put(':id')
   @UseInterceptors(FileFieldsInterceptor([{ name: 'image', maxCount: 1 }]))
   public async update(
     @Param('id') id: UserPublicEntity['id'],
     @Body() updateUserDto: UpdateUserDto,
+    @AuthenticatedUser() authenticatedUser: UserPublicEntity,
     @UploadedFiles()
     @UploadRestrictions([
       {
@@ -73,11 +82,27 @@ export class UserController {
     ])
     files?: UpdateUserUploadedFiles,
   ): Promise<UserPublicEntity> {
+    if (id !== authenticatedUser.id) {
+      throw new ForbiddenException(
+        'This action is forbiden for user. User id must be equal to the authenticated user id',
+      );
+    }
+
     return this.userService.update(id, updateUserDto, files);
   }
 
+  @Auth(JwtAuthGuard)
   @Delete(':id')
-  public async remove(@Param('id') id: UserPublicEntity['id']): Promise<UserPublicEntity> {
+  public async remove(
+    @Param('id') id: UserPublicEntity['id'],
+    @AuthenticatedUser() authenticatedUser: UserPublicEntity,
+  ): Promise<UserPublicEntity> {
+    if (id !== authenticatedUser.id) {
+      throw new ForbiddenException(
+        'This action is forbiden for user. User id must be equal to the authenticated user id',
+      );
+    }
+
     return this.userService.remove(id);
   }
 }
